Minor file refactor
