+++ gui/src/components/Indicators/GenericIndicator.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import '../styles/GenericIndicator.css'; 
+import './styles/GenericIndicator.css'; 
 class GenericIndicator extends React.Component {
   static defaultProps = {
     name: "Un-nammed Indicator",
