+++ gui/src/components/Indicators/styles/GenericIndicator.css
